refactor(engine): cleanup get component stack
