+++ packages/@lwc/engine/src/framework/template.ts
@@ -17,7 +17,7 @@ import {
     isUndefined,
     toString,
 } from '@lwc/shared';
-import { logError } from '../shared/assert';
+import { logError } from '../shared/logger';
 import { VNode, VNodes } from '../3rdparty/snabbdom/types';
 import * as api from './api';
 import { RenderAPI } from './api';
